+++ src/bookshelf/reading-journey/pattern/book-note/book-note-pattern.spec.ts
@@ -2,6 +2,7 @@ import { expect, test } from 'vitest'
 import { bookNotePatterns } from './book-note-pattern'
 
 import { Page } from '../../position/page'
+import { Percentage } from '../../position/percentage'
 
 const validPatterns = {
     started: '{date}: Started',
@@ -27,6 +28,11 @@ test.each([
     ['2025-02-20: Finished', { action: 'finished', date: new Date(2025, 1, 20) }],
     ['2025-02-20: 5-25', { action: 'progress', date: new Date(2025, 1, 20), start: new Page(5), end: new Page(25) }],
     ['2025-02-20: 50', { action: 'progress', date: new Date(2025, 1, 20), start: null, end: new Page(50) }],
+    [
+        '2025-02-20: 2%-30%',
+        { action: 'progress', date: new Date(2025, 1, 20), start: new Percentage(2), end: new Percentage(30) },
+    ],
+    ['2025-02-20: 60%', { action: 'progress', date: new Date(2025, 1, 20), start: null, end: new Percentage(60) }],
 ])('Input "%s" should return matches: %s', (input, expected) => {
     const result = bookNotePatterns(validPatterns, 'yyyy-MM-dd')