+++ src/bookshelf/reading-journey/pattern/daily-note/daily-note-pattern.spec.ts
@@ -2,6 +2,7 @@ import { expect, test } from 'vitest'
 import { dailyNotePatterns } from './daily-note-pattern'
 
 import { Page } from '../../position/page'
+import { Percentage } from '../../position/percentage'
 
 const validPatterns = {
     started: '{book}: Started',
@@ -19,6 +20,11 @@ test.each([
     ['[[Dracula]]: Finished', { action: 'finished', book: '[[Dracula]]' }],
     ['[[Dracula]]: 5-25', { action: 'progress', book: '[[Dracula]]', start: new Page(5), end: new Page(25) }],
     ['[[Dracula]]: 50', { action: 'progress', book: '[[Dracula]]', start: null, end: new Page(50) }],
+    [
+        '[[Dracula]]: 2%-30%',
+        { action: 'progress', book: '[[Dracula]]', start: new Percentage(2), end: new Percentage(30) },
+    ],
+    ['[[Dracula]]: 60%', { action: 'progress', book: '[[Dracula]]', start: null, end: new Percentage(60) }],
 ])('Input "%s" should return matches: %s', (input, expected) => {
     const result = dailyNotePatterns(validPatterns)