+++ src/bookshelf/reading-journey/position/position.spec.ts
@@ -1,11 +1,13 @@
 import { expect, test } from 'vitest'
 import { position } from './position'
 import { Page } from './page'
+import { Percentage } from './percentage'
 
 test.each([
     [1, new Page(1)],
     ['123', new Page(123)],
     [' 123 ', new Page(123)],
+    ['25%', new Percentage(25)],
 ])('position(%o) should return %o', (input, expected) => {
     expect(position(input)).toEqual(expected)
 })